+++ .gitignore
@@ -9,6 +9,7 @@ test
 do-adhoc
 /hardware_testing/pi
 /monitoring/pcap-fun/read-packet
+logs
 
 
 # Prerequisites
