+++ adhoc/adhoc_defines.h
@@ -1,0 +1,11 @@
+#ifndef ADHOC_DEFINES
+#define ADHOC_DEFINES
+
+#include "../defines.h"
+
+
+#define MONITOR_DEBUG       DEBUG && 1
+#define ROUTING_DEBUG       DEBUG && 1
+#define TABLE_DEBUG         DEBUG && 1
+
+#endif // ADHOC_DEFINES
