+++ adhoc/monitor_info.h
@@ -1,3 +1,6 @@
+#ifndef MONITORINFO_H
+#define MONITORINFO_H
+
 /*********************************
  * monitor_info.h
  *
@@ -8,6 +11,7 @@
  ********************************/
 
 #include "../defines.h"
+
 struct local_data
 {
     double batteryLevel;
@@ -18,3 +22,5 @@ struct pair_data
     IP_ADDR pairIp;
     double rss;
 };
+
+#endif
