+++ adhoc/network_monitor.cc
@@ -1,5 +1,4 @@
 #include "network_monitor.h"
-#include "monitor_defines.h"
 
 #include <iostream>
