+++ aodv/aodv.h
@@ -17,7 +17,8 @@
 #include "aodv_rerr.h"
 #include "aodv_params.h"
 
-#include "send_packet.h"
+#include "../adhoc/send_packet.h"
+#include "../adhoc/routing_protocol.h"
 
 #include <vector>
 #include <functional>
