+++ aodv/aodv_defines.h
@@ -5,6 +5,7 @@
 #include <netinet/in.h>
 #include <arpa/inet.h>
 #include "../defines.h"
+#include "../adhoc/adhoc_defines.h"
 
 #define BROADCAST_STR       "255.255.255.255"
 #define HEADER_SIZE         9
@@ -17,8 +18,6 @@
 #define RREQ_DEBUG		    DEBUG && 1 && GLOBAL_AODV_DEBUG
 #define RREP_DEBUG		    DEBUG && 1 && GLOBAL_AODV_DEBUG
 #define RERR_DEBUG          DEBUG && 1 && GLOBAL_AODV_DEBUG
-#define TABLE_DEBUG         DEBUG && 1 && GLOBAL_AODV_DEBUG
-#define MONITOR_DEBUG       DEBUG && 1 && GLOBAL_AODV_DEBUG
-#define ROUTING_DEBUG       DEBUG && 1 && GLOBAL_AODV_DEBUG
+
 
 #endif
