+++ aodv/aodv_routing_table.h
@@ -12,7 +12,7 @@
 
 // include for routing table 
 // TODO: Pull routing table out of RoutingProtocol.h
-#include "routing_protocol.h"
+#include "../adhoc/routing_protocol.h"
 #include "aodv_packets.h"
