+++ aodv/test.cc
@@ -1,5 +1,5 @@
 #include "aodv.h"
-#include "send_packet.h"
+#include "../adhoc/send_packet.h"
 
 #include <assert.h>
 #include "string.h"
