--- logs/192.168.1.0-rtable.txt
@@ -1,3 +1,0 @@
-AODV Log for node 192.168.1.0
-DESTINATION IP : NEXT HOP : TOTAL HOPS : STATUS
-192.168.1.4 : 192.168.1.1 : 0 : ACTIVE
