--- logs/192.168.1.2-rtable.txt
@@ -1,4 +1,0 @@
-AODV Log for node 192.168.1.2
-DESTINATION IP : NEXT HOP : TOTAL HOPS : STATUS
-192.168.1.0 : 192.168.1.0 : 1 : ACTIVE
-192.168.1.4 : 192.168.1.3 : 0 : ACTIVE
