--- logs/192.168.1.3-packet-log.txt
@@ -1,1 +1,0 @@
-Hello node 3! From node 0
