--- logs/192.168.1.3-rtable.txt
@@ -1,4 +1,0 @@
-AODV Log for node 192.168.1.3
-DESTINATION IP : NEXT HOP : TOTAL HOPS : STATUS
-192.168.1.0 : 192.168.1.2 : 2 : ACTIVE
-192.168.1.4 : 192.168.1.4 : 0 : ACTIVE
