--- logs/192.168.1.4-rtable.txt
@@ -1,3 +1,0 @@
-AODV Log for node 192.168.1.4
-DESTINATION IP : NEXT HOP : TOTAL HOPS : STATUS
-192.168.1.0 : 192.168.1.2 : 3 : ACTIVE
