+++ rem/rem.cc
