+++ rem/rem_hardware.h
