+++ rem/rem_model.cc
