+++ rem/rem_sim.cc
